refactor code to improve clarity, begin adding typing
