+++ src/daft/_core.py
@@ -1,18 +1,22 @@
 """Code for Daft"""
 
-__all__ = ["PGM", "Node", "Edge", "Plate"]
+__all__ = ["PGM", ]
 # TODO: should Text be added?
 
-import matplotlib as mpl
 import matplotlib.pyplot as plt
-from matplotlib.patches import Ellipse
-from matplotlib.patches import FancyArrow
-from matplotlib.patches import Rectangle
+from matplotlib.patches import Ellipse, FancyArrow, Rectangle
 
 import numpy as np
 
-from ._exceptions import SameLocationError
-from ._utils import _rendering_context, _pop_multiple
+from typing import Any
+from numpy.typing import NDArray, ArrayLike
+
+
+from .node import Node
+from .edge import Edge
+from .plate import Plate, Text
+from ._utils import _rendering_context
+from ._types import NDArray2, NDArrayF, NDArrayI
 
 # pylint: disable=too-many-arguments, protected-access, unused-argument, too-many-lines
 
@@ -73,24 +77,24 @@ class PGM:
 
     def __init__(
         self,
-        shape=None,
-        origin=None,
-        grid_unit=2.0,
-        node_unit=1.0,
-        observed_style="shaded",
-        alternate_style="inner",
-        line_width=1.0,
-        node_ec="k",
-        node_fc="w",
-        plate_fc="w",
-        directed=True,
-        aspect=1.0,
-        label_params=None,
-        dpi=None,
-    ):
-        self._nodes = {}
-        self._edges = []
-        self._plates = []
+        shape: tuple[float, float] | list[float] | NDArrayF | None = None,
+        origin: ArrayLike | None = None,
+        grid_unit: float = 2.0,
+        node_unit: float = 1.0,
+        observed_style: str = "shaded",
+        alternate_style: str = "inner",
+        line_width: float = 1.0,
+        node_ec: str = "k",
+        node_fc: str = "w",
+        plate_fc: str = "w",
+        directed: bool = True,
+        aspect: float = 1.0,
+        label_params: dict[str, Any] | None = None,
+        dpi: int | None = None,
+    ) -> None:
+        self._nodes: dict[str, Node] = {}
+        self._edges: dict[str, Edge] = []
+        self._plates: dict[str, Plate] = []
         self._dpi = dpi
 
         # if shape and origin are not given, pass a default
@@ -366,7 +370,7 @@ class PGM:
 
         return None
 
-    def render(self, dpi=None):
+    def render(self, dpi: int | None = None) -> plt.Axes:
         """
         Render the :class:`Plate`, :class:`Edge` and :class:`Node` objects in
         the model. This will create a new figure with the correct dimensions
@@ -500,701 +504,3 @@ class PGM:
         if not self.figure:
             self.render()
         self.figure.savefig(fname, *args, **kwargs)
-
-
-class Node:
-    """
-    The representation of a random variable in a :class:`PGM`.
-
-    :param name:
-        The plain-text identifier for the node.
-
-    :param content:
-        The display form of the variable.
-
-    :param x:
-        The x-coordinate of the node in *model units*.
-
-    :param y:
-        The y-coordinate of the node.
-
-    :param scale: (optional)
-        The diameter (or height) of the node measured in multiples of
-        ``node_unit`` as defined by the :class:`PGM` object.
-
-    :param aspect: (optional)
-        The aspect ratio width/height for elliptical nodes; default 1.
-
-    :param observed: (optional)
-        Should this be a conditioned variable?
-
-    :param fixed: (optional)
-        Should this be a fixed (not permitted to vary) variable?
-        If `True`, modifies or over-rides ``diameter``, ``offset``,
-        ``facecolor``, and a few other ``plot_params`` settings.
-        This setting conflicts with ``observed``.
-
-    :param alternate: (optional)
-        Should this use the alternate style?
-
-    :param offset: (optional)
-        The ``(dx, dy)`` offset of the label (in points) from the default
-        centered position.
-
-    :param fontsize: (optional)
-        The fontsize to use.
-
-    :param plot_params: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.patches.Ellipse` constructor.
-
-    :param label_params: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.text.Annotation` constructor. Any kwargs not
-        used by Annontation get passed to :class:`matplotlib.text.Text`.
-
-    :param shape: (optional)
-        String in {ellipse (default), rectangle}
-        If rectangle, aspect and scale holds for rectangle.
-
-    """
-
-    def __init__(
-        self,
-        name,
-        content,
-        x,
-        y,
-        scale=1.0,
-        aspect=None,
-        observed=False,
-        fixed=False,
-        alternate=False,
-        offset=(0.0, 0.0),
-        fontsize=None,
-        plot_params=None,
-        label_params=None,
-        shape="ellipse",
-    ):
-        # Check Node style.
-        # Iterable is consumed, so first condition checks if two or more are
-        # true
-        node_style = iter((observed, alternate, fixed))
-        if not (
-            (any(node_style) and not any(node_style))
-            or not any((observed, alternate, fixed))
-        ):
-            msg = "A node cannot be more than one of `observed`, `fixed`, or `alternate`."
-            raise ValueError(msg)
-
-        self.observed = observed
-        self.fixed = fixed
-        self.alternate = alternate
-
-        # Metadata.
-        self.name = name
-        self.content = content
-
-        # Coordinates and dimensions.
-        self.x, self.y = float(x), float(y)
-        self.scale = float(scale)
-        if self.fixed:
-            self.scale /= 6.0
-        if aspect is not None:
-            self.aspect = float(aspect)
-        else:
-            self.aspect = aspect
-
-        # Set fontsize
-        self.fontsize = fontsize if fontsize else mpl.rcParams["font.size"]
-
-        # Display parameters.
-        self.plot_params = dict(plot_params) if plot_params else {}
-
-        # Text parameters.
-        self.offset = list(offset)
-        self.label_params = dict(label_params) if label_params else None
-
-        # Shape
-        if shape in ["ellipse", "rectangle"]:
-            self.shape = shape
-        else:
-            print("Warning: wrong shape value, set to ellipse instead")
-            self.shape = "ellipse"
-
-    def render(self, ctx):
-        """
-        Render the node.
-
-        :param ctx:
-            The :class:`_rendering_context` object.
-
-        """
-        # Get the axes and default plotting parameters from the rendering
-        # context.
-        ax = ctx.ax()
-
-        # Resolve the plotting parameters.
-        plot_params = dict(self.plot_params)
-
-        plot_params["lw"] = _pop_multiple(
-            plot_params, ctx.line_width, "lw", "linewidth"
-        )
-
-        plot_params["ec"] = plot_params["edgecolor"] = _pop_multiple(
-            plot_params, ctx.node_ec, "ec", "edgecolor"
-        )
-
-        fc_is_set = "fc" in plot_params or "facecolor" in plot_params
-        plot_params["fc"] = _pop_multiple(
-            plot_params, ctx.node_fc, "fc", "facecolor"
-        )
-        fc = plot_params["fc"]
-
-        plot_params["alpha"] = plot_params.get("alpha", 1)
-
-        # And the label parameters.
-        if self.label_params is None:
-            label_params = dict(ctx.label_params)
-        else:
-            label_params = dict(self.label_params)
-
-        label_params["va"] = _pop_multiple(
-            label_params, "center", "va", "verticalalignment"
-        )
-
-        label_params["ha"] = _pop_multiple(
-            label_params, "center", "ha", "horizontalalignment"
-        )
-
-        # Deal with ``fixed`` nodes.
-        scale = self.scale
-        if self.fixed:
-            # MAGIC: These magic numbers should depend on the grid/node units.
-            self.offset[1] += 6
-
-            label_params["va"] = "baseline"
-            label_params.pop("verticalalignment", None)
-            label_params.pop("ma", None)
-
-            if not fc_is_set:
-                plot_params["fc"] = "k"
-
-        diameter = ctx.node_unit * scale
-        if self.aspect is not None:
-            aspect = self.aspect
-        else:
-            aspect = ctx.aspect
-
-        # Set up an observed node or alternate node. Note the fc INSANITY.
-        if self.observed and not self.fixed:
-            style = ctx.observed_style
-        elif self.alternate and not self.fixed:
-            style = ctx.alternate_style
-        else:
-            style = False
-
-        if style:
-            # Update the plotting parameters depending on the style of
-            # observed node.
-            h = float(diameter)
-            w = aspect * float(diameter)
-            if style == "shaded":
-                plot_params["fc"] = "0.7"
-            elif style == "outer":
-                h = diameter + 0.1 * diameter
-                w = aspect * diameter + 0.1 * diameter
-            elif style == "inner":
-                h = diameter - 0.1 * diameter
-                w = aspect * diameter - 0.1 * diameter
-                plot_params["fc"] = fc
-
-            # Draw the background ellipse.
-            if self.shape == "ellipse":
-                bg = Ellipse(
-                    xy=ctx.convert(self.x, self.y),
-                    width=w,
-                    height=h,
-                    **plot_params,
-                )
-            elif self.shape == "rectangle":
-                # Adapt to make Rectangle the same api than ellipse
-                wi = w
-                xy = ctx.convert(self.x, self.y)
-                xy[0] = xy[0] - wi / 2.0
-                xy[1] = xy[1] - h / 2.0
-
-                bg = Rectangle(xy=xy, width=wi, height=h, **plot_params)
-            else:
-                # Should never append
-                raise (
-                    ValueError(
-                        "Wrong shape in object causes an error in render"
-                    )
-                )
-
-            ax.add_artist(bg)
-
-            # Reset the face color.
-            plot_params["fc"] = fc
-
-        # Draw the foreground ellipse.
-        if not fc_is_set and not self.fixed and self.observed:
-            plot_params["fc"] = "none"
-
-        if self.shape == "ellipse":
-            el = Ellipse(
-                xy=ctx.convert(self.x, self.y),
-                width=diameter * aspect,
-                height=diameter,
-                **plot_params,
-            )
-        elif self.shape == "rectangle":
-            # Adapt to make Rectangle the same api than ellipse
-            wi = diameter * aspect
-            xy = ctx.convert(self.x, self.y)
-            xy[0] = xy[0] - wi / 2.0
-            xy[1] = xy[1] - diameter / 2.0
-
-            el = Rectangle(xy=xy, width=wi, height=diameter, **plot_params)
-        else:
-            # Should never append
-            raise (
-                ValueError("Wrong shape in object causes an error in render")
-            )
-
-        ax.add_artist(el)
-
-        # Reset the face color.
-        plot_params["fc"] = fc
-
-        # Annotate the node.
-        ax.annotate(
-            self.content,
-            ctx.convert(self.x, self.y),
-            xycoords="data",
-            xytext=self.offset,
-            textcoords="offset points",
-            size=self.fontsize,
-            **label_params,
-        )
-
-        return el
-
-    def get_frontier_coord(self, target_xy, ctx, edge):
-        """
-        Get the coordinates of the point of intersection between the
-        shape of the node and a line starting from the center of the node to an
-        arbitrary point. Will throw a :class:`SameLocationError` if the nodes
-        contain the same `x` and `y` coordinates. See the example of rectangle
-        below:
-
-        .. code-block:: python
-
-            _____________
-            |            |    ____--X (target_node)
-            |        __--X----
-            |     X--    |(return coordinate of this point)
-            |            |
-            |____________|
-
-        :target_xy: (x float, y float)
-            A tuple of coordinate of target node
-
-        """
-
-        # Scale the coordinates appropriately.
-        x1, y1 = ctx.convert(self.x, self.y)
-        x2, y2 = target_xy[0], target_xy[1]
-
-        # Aspect ratios.
-        if self.aspect is not None:
-            aspect = self.aspect
-        else:
-            aspect = ctx.aspect
-
-        if self.shape == "ellipse":
-            # Compute the distances.
-            dx, dy = x2 - x1, y2 - y1
-            if dx == 0.0 and dy == 0.0:
-                raise SameLocationError(edge)
-            dist1 = np.sqrt(dy * dy + dx * dx / float(aspect * aspect))
-
-            # Compute the fractional effect of the radii of the nodes.
-            alpha1 = 0.5 * ctx.node_unit * self.scale / dist1
-
-            # Get the coordinates of the starting position.
-            x0, y0 = x1 + alpha1 * dx, y1 + alpha1 * dy
-
-            return x0, y0
-
-        elif self.shape == "rectangle":
-            dx, dy = x2 - x1, y2 - y1
-
-            # theta = np.angle(complex(dx, dy))
-            # print(theta)
-            # left or right intersection
-            dxx1 = self.scale * aspect / 2.0 * (np.sign(dx) or 1.0)
-            dyy1 = (
-                self.scale
-                * aspect
-                / 2.0
-                * np.abs(dy / dx)
-                * (np.sign(dy) or 1.0)
-            )
-            val1 = np.abs(complex(dxx1, dyy1))
-
-            # up or bottom intersection
-            dxx2 = self.scale * 0.5 * np.abs(dx / dy) * (np.sign(dx) or 1.0)
-            dyy2 = self.scale * 0.5 * (np.sign(dy) or 1.0)
-            val2 = np.abs(complex(dxx2, dyy2))
-
-            if val1 < val2:
-                return x1 + dxx1, y1 + dyy1
-            else:
-                return x1 + dxx2, y1 + dyy2
-
-        else:
-            # Should never append
-            raise ValueError("Wrong shape in object causes an error")
-
-
-class Edge:
-    """
-    An edge between two :class:`Node` objects.
-
-    :param node1:
-        The first :class:`Node`.
-
-    :param node2:
-        The second :class:`Node`. The arrow will point towards this node.
-
-    :param directed: (optional)
-        Should the edge be directed from ``node1`` to ``node2``? In other
-        words: should it have an arrow?
-
-    :param label: (optional)
-        A string to annotate the edge.
-
-    :param xoffset: (optional)
-        The x-offset from the middle of the arrow to plot the label.
-        Only takes effect if `label` is defined in `plot_params`.
-
-    :param yoffset: (optional)
-        The y-offset from the middle of the arrow to plot the label.
-        Only takes effect if `label` is defined in `plot_params`.
-
-    :param plot_params: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.patches.FancyArrow` constructor to adjust
-        edge behavior.
-
-    :param label_params: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.axes.Axes.annotate` constructor to adjust
-        label behavior.
-
-    """
-
-    def __init__(
-        self,
-        node1,
-        node2,
-        directed=True,
-        label=None,
-        xoffset=0,
-        yoffset=0.1,
-        plot_params=None,
-        label_params=None,
-    ):
-        self.node1 = node1
-        self.node2 = node2
-        self.directed = directed
-        self.label = label
-        self.xoffset = xoffset
-        self.yoffset = yoffset
-        self.plot_params = dict(plot_params) if plot_params else {}
-        self.label_params = dict(label_params) if label_params else {}
-
-    def _get_coords(self, ctx):
-        """
-        Get the coordinates of the line.
-
-        :param conv:
-            A callable coordinate conversion.
-
-        :returns:
-            * ``x0``, ``y0``: the coordinates of the start of the line.
-            * ``dx0``, ``dy0``: the displacement vector.
-
-        """
-        # Scale the coordinates appropriately.
-        x1, y1 = ctx.convert(self.node1.x, self.node1.y)
-        x2, y2 = ctx.convert(self.node2.x, self.node2.y)
-
-        x3, y3 = self.node1.get_frontier_coord((x2, y2), ctx, self)
-        x4, y4 = self.node2.get_frontier_coord((x1, y1), ctx, self)
-
-        return x3, y3, x4 - x3, y4 - y3
-
-    def render(self, ctx):
-        """
-        Render the edge in the given axes.
-
-        :param ctx:
-            The :class:`_rendering_context` object.
-
-        """
-        ax = ctx.ax()
-
-        plot_params = self.plot_params
-        plot_params["linewidth"] = _pop_multiple(
-            plot_params, ctx.line_width, "lw", "linewidth"
-        )
-
-        plot_params["linestyle"] = plot_params.get("linestyle", "-")
-
-        # Add edge annotation.
-        if self.label is not None:
-            x, y, dx, dy = self._get_coords(ctx)
-            ax.annotate(
-                self.label,
-                [x + 0.5 * dx + self.xoffset, y + 0.5 * dy + self.yoffset],
-                xycoords="data",
-                xytext=[0, 3],
-                textcoords="offset points",
-                ha="center",
-                va="center",
-                **self.label_params,
-            )
-
-        if self.directed:
-            plot_params["ec"] = _pop_multiple(
-                plot_params, "k", "ec", "edgecolor"
-            )
-            plot_params["fc"] = _pop_multiple(
-                plot_params, "k", "fc", "facecolor"
-            )
-            plot_params["head_length"] = plot_params.get("head_length", 0.25)
-            plot_params["head_width"] = plot_params.get("head_width", 0.1)
-
-            # Build an arrow.
-            args = self._get_coords(ctx)
-
-            # zero lengh arrow produce error
-            if not (args[2] == 0.0 and args[3] == 0.0):
-                ar = FancyArrow(
-                    *self._get_coords(ctx),
-                    width=0,
-                    length_includes_head=True,
-                    **plot_params,
-                )
-
-                # Add the arrow to the axes.
-                ax.add_artist(ar)
-                return ar
-
-            else:
-                print(args[2], args[3])
-
-        else:
-            plot_params["color"] = plot_params.get("color", "k")
-
-            # Get the right coordinates.
-            x, y, dx, dy = self._get_coords(ctx)
-
-            # Plot the line.
-            line = ax.plot([x, x + dx], [y, y + dy], **plot_params)
-            return line
-
-
-class Plate:
-    """
-    A plate to encapsulate repeated independent processes in the model.
-
-    :param rect:
-        The rectangle describing the plate bounds in model coordinates.
-        This is [x-start, y-start, x-length, y-length].
-
-    :param label: (optional)
-        A string to annotate the plate.
-
-    :param label_offset: (optional)
-        The x- and y- offsets of the label text measured in points.
-
-    :param shift: (optional)
-        The vertical "shift" of the plate measured in model units. This will
-        move the bottom of the panel by ``shift`` units.
-
-    :param position: (optional)
-        One of ``"{vertical} {horizontal}"`` where vertical is ``"bottom"``
-        or ``"middle"`` or ``"top"`` and horizontal is ``"left"`` or
-        ``"center"`` or ``"right"``.
-
-    :param fontsize: (optional)
-        The fontsize to use.
-
-    :param rect_params: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.patches.Rectangle` constructor, which defines
-        the properties of the plate.
-
-    :param bbox: (optional)
-        A dictionary of parameters to pass to the
-        :class:`matplotlib.axes.Axes.annotate` constructor, which defines
-        the box drawn around the text.
-
-    """
-
-    def __init__(
-        self,
-        rect,
-        label=None,
-        label_offset=(5, 5),
-        shift=0,
-        position="bottom left",
-        fontsize=None,
-        rect_params=None,
-        bbox=None,
-    ):
-        self.rect = rect
-        self.label = label
-        self.label_offset = label_offset
-        self.shift = shift
-
-        if fontsize is not None:
-            self.fontsize = fontsize
-        else:
-            self.fontsize = mpl.rcParams["font.size"]
-
-        if rect_params is not None:
-            self.rect_params = dict(rect_params)
-        else:
-            self.rect_params = None
-
-        if bbox is not None:
-            self.bbox = dict(bbox)
-
-            # Set the awful default blue color to transparent
-            if "fc" not in self.bbox.keys():
-                self.bbox["fc"] = "none"
-        else:
-            self.bbox = None
-
-        self.position = position
-
-    def render(self, ctx):
-        """
-        Render the plate in the given axes.
-
-        :param ctx:
-            The :class:`_rendering_context` object.
-
-        """
-        ax = ctx.ax()
-
-        shift = np.array([0, self.shift], dtype=np.float64)
-        rect = np.atleast_1d(self.rect)
-        bottom_left = ctx.convert(*(rect[:2] + shift))
-        top_right = ctx.convert(*(rect[:2] + rect[2:]))
-        rect = np.concatenate([bottom_left, top_right - bottom_left])
-
-        if self.rect_params is not None:
-            rect_params = self.rect_params
-        else:
-            rect_params = {}
-
-        rect_params["ec"] = _pop_multiple(rect_params, "k", "ec", "edgecolor")
-        rect_params["fc"] = _pop_multiple(
-            rect_params, ctx.plate_fc, "fc", "facecolor"
-        )
-        rect_params["lw"] = _pop_multiple(
-            rect_params, ctx.line_width, "lw", "linewidth"
-        )
-        rectangle = Rectangle(rect[:2], *rect[2:], **rect_params)
-
-        ax.add_artist(rectangle)
-
-        if self.label is not None:
-            offset = np.array(self.label_offset, dtype=np.float64)
-            if "left" in self.position:
-                position = rect[:2]
-                ha = "left"
-            elif "right" in self.position:
-                position = rect[:2]
-                position[0] += rect[2]
-                ha = "right"
-                offset[0] = -offset[0]
-            elif "center" in self.position:
-                position = rect[:2]
-                position[0] = rect[2] / 2 + rect[0]
-                ha = "center"
-            else:
-                raise RuntimeError(
-                    f"Unknown positioning string: {self.position}"
-                )
-
-            if "bottom" in self.position:
-                va = "bottom"
-            elif "top" in self.position:
-                position[1] = rect[1] + rect[3]
-                offset[1] = -offset[1] - 0.1
-                va = "top"
-            elif "middle" in self.position:
-                position[1] += rect[3] / 2
-                va = "center"
-            else:
-                raise RuntimeError(
-                    f"Unknown positioning string: {self.position}"
-                )
-
-            ax.annotate(
-                self.label,
-                xy=position,
-                xycoords="data",
-                xytext=offset,
-                textcoords="offset points",
-                size=self.fontsize,
-                bbox=self.bbox,
-                horizontalalignment=ha,
-                verticalalignment=va,
-            )
-
-        return rectangle
-
-
-class Text(Plate):
-    """
-    A subclass of plate to writing text using grid coordinates. Any **kwargs
-    are passed through to :class:`PGM.Plate`.
-
-    :param x:
-        The x-coordinate of the text in *model units*.
-
-    :param y:
-        The y-coordinate of the text.
-
-    :param label:
-        A string to write.
-
-    :param fontsize: (optional)
-        The fontsize to use.
-
-    """
-
-    def __init__(self, x, y, label, fontsize=None):
-        self.rect = [x, y, 0.0, 0.0]
-        self.label = label
-        self.fontsize = fontsize
-        self.label_offset = [0.0, 0.0]
-        self.bbox = {"fc": "none", "ec": "none"}
-        self.rect_params = {"ec": "none"}
-
-        super().__init__(
-            rect=self.rect,
-            label=self.label,
-            label_offset=self.label_offset,
-            fontsize=self.fontsize,
-            rect_params=self.rect_params,
-            bbox=self.bbox,
-        )
